extract classes to files for readability
